requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.7
pinecone-client==2.2.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23